// Client/employee-tracker/src/types/contract.d.ts
export interface Contract {
    id: number;
    first_name: string;
    middle_name?: string;
    last_name: string;
    start_date: string;
    end_date: string;
    project: string;
    line_manager: string;
    phone_number: string;
    email: string;
    email2?: string;
    email3?: string;
    email4?: string;
  }